ACM question in POJ, No.3662
http://poj.org/problem?id=3662

Description

Farmer John wants to set up a telephone line at his farm. Unfortunately,
the phone company is uncooperative, so he needs to pay for some of the
cables required to connect his farm to the phone system.

There are N (1 ≤ N ≤ 1,000) forlorn telephone poles conveniently
numbered 1..N that are scattered around Farmer John's property; no
cables connect any them. A total of P (1 ≤ P ≤ 10,000) pairs of poles
can be connected by a cable; the rest are too far apart.

The i-th cable can connect the two distinct poles Ai and Bi, with length
Li (1 ≤ Li ≤ 1,000,000) units if used. The input data set never names
any {Ai, Bi} pair more than once. Pole 1 is already connected to the
phone system, and pole N is at the farm. Poles 1 and N need to be
connected by a path of cables; the rest of the poles might be used or
might not be used.

As it turns out, the phone company is willing to provide Farmer John
with K (0 ≤ K < N) lengths of cable for free. Beyond that he will have
to pay a price equal to the length of the longest remaining cable he
requires (each pair of poles is connected with a separate cable), or 0
if he does not need any additional cables.

Determine the minimum amount that Farmer John must pay.

Input

* Line 1: Three space-separated integers: N, P, and K
* Lines 2..P+1: Line i+1 contains the three space-separated integers:
Ai, Bi, and Li

Output

* Line 1: A single integer, the minimum amount Farmer John can pay. If
it is impossible to connect the farm to the phone company, print -1.

Sample Input

5 7 1
1 2 5
3 1 4
2 4 8
3 2 3
5 2 9
3 4 7
4 5 6
Sample Output

4